Move indexer to indexer/code
